+++ pages/components/trainer_prompts.py
@@ -48,11 +48,34 @@ class TrainerPrompts():
         - The user can also ask about the current exercise, such as what muscles it works out, how to do it, etc. The trainer will respond with the answer.
         '''
         return self.get_prompt(
-            input_variables=['history', 'input', 'workput_routine'],
+            input_variables=['history', 'input'],
             template=
             '''
-            You are a personal trainer and you are talking to a client. You already have a workout routine for the client: {workout_routine}. You will tell the client what to do next based on the client's input.\n\nCurrent conversation:\n{history}\n\nClient: {input}\nTrainer:
-            '''    
+            You are a personal trainer conversing with a client. Your role is to guide the client through a workout routine, exercise by exercise. Follow these instructions:
+
+            1. Start by asking the client to select a workout routine. The client will give you an ordered list. Do not proceed until the client specifies one. The first entry 0. is notes about the workout routine. 
+            2. Once a routine is selected, describe the first exercise. Be consise. 
+            3. Wait for the client to say they have completed the current exercise before describing the next one. 
+            The client must explicitly state they have finished an exercise (e.g., 'I have finished [exercise name]' or 'I am done') before you move on.
+            4. Only if the client asks questions about an exercise, provide clear and concise answers.
+            5. Continue this process until all exercises in the routine are completed.
+            6. After the last exercise is completed, end the session with '[Finished Workout]'.
+
+            Remember:
+            - Be patient and keep asking for a workout routine if the client hasn't specified one.
+            - Ensure safety and clarity in exercise instructions.
+            - Respond promptly to client's completion of exercises and questions.
+            \n\nCurrent conversation:\n{history}\n\Question: {input}\nAnswer:
+            '''
+            
+            
+            
+            
+            
+            
+            # '''
+            # You are a personal trainer and you are talking to a client. The client will begin by saying I would like to select a workout routine. You are to ask the client for a workout routine. Only after the client gives you a workout routine, you are to guide the client through the workout routine exercise by exercise. Only start the routine if the client has given you one, otherwise keep asking client to select one. Provide the next exercise when the client says they have completed the current exercise. The client may ask questions about each exercise. When all exercises are complete, add "[Finished Workout]" to the end. \n\nCurrent conversation:\n{history}\n\nQuestion: {input}\nAnswer:
+            # '''    
         )
     
     def trainer_new_session_chatmodel_prompt(self):
@@ -63,9 +86,25 @@ class TrainerPrompts():
             input_variables=['history', 'input'],
             template=
             '''
-            You are a personal trainer and you are talking to a client. You will start to record a new workout routine for the client. The client will start by asking to start a session. You will say, "Go ahead". The client will tell you what exercise they just completed. You will maintain a list of exercises performed by the client. The client may also ask how to do a workout, only if they ask, give a brief explanation of that workout. When the client tells you "I'm done" or "Finished" or something similar, you will respond with the list of exercises performed by the client and add "[Finished Workout]" at the end.\n\nCurrent conversation:\n{history}\n\nClient: {input}\nTrainer:
+            You are a personal trainer and you are talking to a client. You will start to record a new workout routine for the client. The client will start by asking to start a session. You will say, "Go ahead". The client will tell you what exercise they just completed, always wait for the user. You will maintain a list of exercises performed by the client. The client may also ask how to do a workout. ONLY and WHEN they ask, give a brief explanation of that workout. When the client tells you "I'm done with my workout" or "I am finished with my workout" or something similar, you will respond with the list of exercises performed by the client and add "[Finished Workout]" at the end.\n\nCurrent conversation:\n{history}\n\Question: {input}\nAnswer:
             '''
         )
+        
+        ''' Example conversation:
+        Client: I want to start a workout session.
+        Trainer: Go ahead.
+        Client: I did 10 pushups.
+        Trainer: Great job! Keep going.
+        Client: I did 10 situps.
+        Trainer: Awesome! You are doing well.
+        Client: How do I do a pullup?
+        Trainer: A pullup is done like this: ...
+        Client: I did 10 pullups.
+        Trainer: Great job! Keep going.
+        Client: I'm done.
+        Trainer: You did 10 pushups, 10 situps, 10 pullups. [Finished Workout]
+        '''
+        
         
 
     def default_llama27b_prompt(self):
