--- test/load_pdfs.py
@@ -1,4 +1,0 @@
-import os
-
-pdfs = os.path.expanduser("/home/ftpuser/")
-print(os.listdir(pdfs))
